swc_atoms
Type Alias JsWord

pub type JsWord = Atom<JsWordStaticSet>;

JsWord is an interned string.

This type should be used instead of String for values, because lots of values are duplicated. For example, if an identifier is named myVariable, there will be lots of identifier usages with the value myVariable.

This type

makes equality comparison faster.

reduces memory usage.

Aliased Type

struct JsWord { /* private fields */ }