Crate swc_atoms

See Atom and UnsafeAtom

Macros

atom

Creates an Atom from a constant.

lazy_atom

Creates an Atom from a constant.

Structs

Atom

Clone-on-write string.

AtomStore

AtomStoreCell

A fast internally mutable cell for AtomStore.

UnsafeAtom

UnsafeAtom is a wrapper around Atom that does not allocate, but extremely unsafe.

Type Aliases

StaticString

This should be used as a key for hash maps and hash sets.